backend/requirements.txt
astroid==2.4.2
colorama==0.4.3
Django==2.2
django-grappelli==2.13.4
django-guardian==2.3.0
django-python3-ldap==0.11.3
django-rest-framework==0.1.0
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
isort==5.5.2
lazy-object-proxy==1.4.3
ldap3==2.8.1
mccabe==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
